+++ rasa_nlu/server.py
@@ -98,7 +98,7 @@ def create_app(config, component_builder=None):
     def train():
         data_string = request.get_data(as_text=True)
         current_app.data_router.start_train_process(data_string, request.args)
-        return jsonify(info="training started. Current pids: {}".format(current_app.data_router.train_procs))
+        return jsonify(info="training started.", training_process_ids=current_app.data_router.train_proc_ids())
 
     logging.basicConfig(filename=config['log_file'], level=config['log_level'])
     logging.captureWarnings(True)